Fix i18n review notes and hook gates
Document the shipped UI string message mappings and extend catalog
placeholder coverage for error templates.

Resolve the post-turn hook failures by making Makefile tool defaults robust
when hook PATH entries omit Cargo or Bun, and clear the lint findings that
blocked `make lint`.
